py-code/requirements.txt
drs
datasets
guesslang-experimental
